requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-Compress==1.13
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
